Executing md-babel should only work in code blocks (#20)
I installed the ESM only packages [unist-util-find] and
[mdast-util-from-markdown]. The packages are used to determine whether a
line in a Markdown document is in a code block.
[mdast-util-from-markdown] transforms a Markdown document to an AST.
[unist-util-find] traverses the tree to find a satisfying node. This is
done in `isInCodeBlock`. The function is then used to determine if the
position in the document is in a code block. Inline code blocks are
ignored. So, in `simple.md` `md-babel` is not executed when the cursor
in `echo "Inline Code"`.

The tests only test whether I glued the packages correctly together.
Most of the work is actually done by them.

Because the packages are ESM only, I had to change the settings in
`package.json` and `tsconfig.json`.

Also, I set the default formatter to the Prettier Extension in the
project settings of VS Code.

I did not have to use a plugin for GitHub flavoured Markdown because
fenced code blocks (created with backticks `` ` `` or tildes `~`) are
already in the CommonMark standard.

Related to: https://github.com/md-babel/vscode-md-babel/issues/16
Closes: https://github.com/md-babel/vscode-md-babel/issues/10

See also: https://stackoverflow.com/a/62554884
See also: https://stackoverflow.com/questions/77427051/how-do-i-disable-es-module-errors-in-vscode
See also: https://www.totaltypescript.com/tsconfig-cheat-sheet
See also: https://www.typescriptlang.org/docs/handbook/modules/theory.html#the-module-output-format
See also: https://www.markdownguide.org/extended-syntax/#fenced-code-blocks

[unist-util-find]: https://github.com/syntax-tree/unist-util-find
[mdast-util-from-markdown]: https://github.com/syntax-tree/mdast-util-from-markdown