Use md-babel in path if executablePath is not set
If the setting mdBabel.executablePath is not set, the extension tries to
find md-babel in the path environment variable. This is done with the
npm package which. If md-babel is also not found in $PATH, it shows an
error message in VS Code.

I changed the error message to also indicate that md-babel was not found
in $PATH.

See also: https://www.npmjs.com/package/which