Find md-babel executable via $PATH (#23)
If the setting mdBabel.executablePath is not set, the extension tries to
find md-babel in the path environment variable. This is done with the
npm package which, which was installed in this PR. If md-babel is also
not found in $PATH, it shows an error message in VS Code.

I changed the error message to also indicate that md-babel was not found
in $PATH. The message is show at the right bottom corner of VS Code.

which was chosen because it is maintained by npm and often used by other
projects.

For more details, read the discussion in the PR.

See also: https://www.npmjs.com/package/which
Closes: https://github.com/md-babel/vscode-md-babel/issues/2
See pull request: https://github.com/md-babel/vscode-md-babel/pull/23