Check if the current OS is supported by md-babel (#25)
If the OS is supported, nothing happens.

If the OS is not supported, an error message is shown on the first
activation of the extension. The error message tells the user that
currently, only Linux and macOS are supported by md-babel. Whether the
extension was already activated once, is stored in the global state of
the extension.

An OS is supported if `os.platform()` return `'linux'` or `'darwin'`.

For more details, for example a screenshot of the error message, see the
discussion in the PR.

See also: https://nodejs.org/api/os.html#osplatform
See pull request: https://github.com/md-babel/vscode-md-babel/pull/25